backend/app/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
python-dotenv==1.0.1
reportlab==4.1.0
scapy==2.5.0
volatility3==2.5.0
celery==5.3.6
dotenv
python-magic==0.4.27
attrs==23.2.0
aiohttp==3.9.3
cloudinary==1.37.0
python-multipart==0.0.9
fastapi-limiter==0.1.6
redis==5.0.1
passlib==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
pydantic==2.6.1
pydantic-settings==2.1.0
python-magic-bin==0.4.14; sys_platform == 'win32'
itsdangerous==2.1.2
starlette>=0.27.0
python-json-logger>=2.0.7
elasticsearch[async]==8.12.1
aioredis
